+++ flutter_google_places_sdk_android/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.1.2+5
+
+* Updating interface version to 0.2.4+3
+* Implementing `updateSettings` interface method
+
+## 0.1.2+4
+
+* Allowing android locale to pass null country (will use current locale country as default)
+
 ## 0.1.2+3
 
 * replacing jcenter for mavenCentral