+++ flutter_google_places_sdk_ios/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.2+3
+
+* Updating platform interface to 0.2.4+3
+
+## 0.1.2+2
+
+* Using locationBias and locationRestriction
+
 ## 0.1.2+1
 
 * Send viewport parameter